+++ ktlint-cli/src/test/kotlin/com/pinterest/ktlint/cli/api/BaselineCLITest.kt
@@ -21,7 +21,7 @@ class BaselineCLITest {
                 projectName,
                 listOf(
                     "TestBaselineFile.kt.test",
-                    "some/path/to/TestBaselineFile2.kt.test",
+                    "some/path/to/TestBaselineFile.kt.test",
                 ),
             ) {
                 SoftAssertions()
@@ -29,11 +29,7 @@ class BaselineCLITest {
                         assertErrorExitCode()
                         assertThat(normalOutput)
                             .containsLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
-                            .containsLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"))
-                            .containsLineMatching(Regex(".*/$projectName/some/path/to/TestBaselineFile2.kt.test:1:25: Unnecessary block.*"))
-                            .containsLineMatching(
-                                Regex(".*/$projectName/some/path/to/TestBaselineFile2.kt.test:2:1: Unexpected blank line.*"),
-                            )
+                            .containsLineMatching(Regex(".*/$projectName/some/path/to/TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
                     }.assertAll()
             }
     }
@@ -54,7 +50,7 @@ class BaselineCLITest {
                     listOf(
                         "--baseline=$baselinePath",
                         "TestBaselineFile.kt.test",
-                        "some/path/to/TestBaselineFile2.kt.test",
+                        "some/path/to/TestBaselineFile.kt.test",
                     ),
                 ) {
                     SoftAssertions()
@@ -62,14 +58,11 @@ class BaselineCLITest {
                             assertNormalExitCode()
                             assertThat(normalOutput)
                                 .doesNotContainLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
-                                .doesNotContainLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"))
                                 .doesNotContainLineMatching(
                                     Regex(".*/$projectName/some/path/to/TestBaselineFile.kt.test:1:24: Unnecessary block.*"),
-                                ).doesNotContainLineMatching(
-                                    Regex(".*/$projectName/some/path/to/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"),
                                 ).containsLineMatching(
                                     Regex(
-                                        ".*Baseline file '$baselinePath' contains 6 reference\\(s\\) to rule ids without a rule set id. " +
+                                        ".*Baseline file '$baselinePath' contains 4 reference\\(s\\) to rule ids without a rule set id. " +
                                             "For those references the rule set id 'standard' is assumed. It is advised to regenerate " +
                                             "this baseline file.*",
                                     ),
@@ -92,7 +85,7 @@ class BaselineCLITest {
                     listOf(
                         "--baseline=$baselinePath",
                         "TestBaselineFile.kt.test",
-                        "some/path/to/TestBaselineFile2.kt.test",
+                        "some/path/to/TestBaselineFile.kt.test",
                     ),
                 ) {
                     SoftAssertions()
@@ -100,14 +93,11 @@ class BaselineCLITest {
                             assertNormalExitCode()
                             assertThat(normalOutput)
                                 .doesNotContainLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
-                                .doesNotContainLineMatching(Regex(".*/$projectName/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"))
                                 .doesNotContainLineMatching(
                                     Regex(".*/$projectName/some/path/to/TestBaselineFile.kt.test:1:24: Unnecessary block.*"),
-                                ).doesNotContainLineMatching(
-                                    Regex(".*/$projectName/some/path/to/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"),
                                 ).containsLineMatching(
                                     Regex(
-                                        ".*Baseline file '$baselinePath' contains 6 reference\\(s\\) to rule ids without a rule set id. " +
+                                        ".*Baseline file '$baselinePath' contains 4 reference\\(s\\) to rule ids without a rule set id. " +
                                             "For those references the rule set id 'standard' is assumed. It is advised to regenerate " +
                                             "this baseline file.*",
                                     ),
@@ -129,7 +119,7 @@ class BaselineCLITest {
                     listOf(
                         "--baseline=$baselinePath",
                         "TestBaselineFile.kt.test",
-                        "some/path/to/TestBaselineFile2.kt.test",
+                        "some/path/to/TestBaselineFile.kt.test",
                     ),
                 ) {
                     SoftAssertions()
@@ -137,13 +127,11 @@ class BaselineCLITest {
                             assertNormalExitCode()
                             assertThat(normalOutput)
                                 .doesNotContainLineMatching(Regex("^TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
-                                .doesNotContainLineMatching(Regex("^TestBaselineFile.kt.test:2:1: Unexpected blank line.*"))
                                 .doesNotContainLineMatching(Regex("^some/path/to/TestBaselineFile.kt.test:1:24: Unnecessary block.*"))
-                                .doesNotContainLineMatching(Regex("^some/path/to/TestBaselineFile.kt.test:2:1: Unexpected blank line.*"))
                                 .containsLineMatching(
                                     Regex(
                                         // Escape "\" in baseline path for Windows
-                                        ".*Baseline file '${baselinePath.replace("\\", "\\\\")}' contains 6 " +
+                                        ".*Baseline file '${baselinePath.replace("\\", "\\\\")}' contains 4 " +
                                             "reference\\(s\\) to rule ids without a rule set id. For those references the rule set id " +
                                             "'standard' is assumed. It is advised to regenerate this baseline file.*",
                                     ),
@@ -167,19 +155,26 @@ class BaselineCLITest {
                 listOf(
                     "--baseline=$baselinePath",
                     "TestBaselineExtraErrorFile.kt.test",
-                    "some/path/to/TestBaselineExtraErrorFile2.kt.test",
+                    "some/path/to/TestBaselineExtraErrorFile.kt.test",
                 ),
             ) {
                 SoftAssertions()
                     .apply {
                         assertErrorExitCode()
                         assertThat(normalOutput)
-                            .containsLineMatching(Regex(".*/$projectName/TestBaselineExtraErrorFile.kt.test:2:1: Unexpected blank line.*"))
                             .containsLineMatching(
-                                Regex(".*/$projectName/some/path/to/TestBaselineExtraErrorFile2.kt.test:2:1: Unexpected blank line.*"),
+                                Regex(
+                                    ".*/$projectName/TestBaselineExtraErrorFile.kt.test:1:1: Replace the block comment with an " +
+                                        "EOL comment.*",
+                                ),
                             ).containsLineMatching(
                                 Regex(
-                                    ".*Baseline file '$baselinePath' contains 6 reference\\(s\\) to rule ids without a rule set id. For " +
+                                    ".*/$projectName/some/path/to/TestBaselineExtraErrorFile.kt.test:1:1: Replace the block comment with " +
+                                        "an EOL comment.*",
+                                ),
+                            ).containsLineMatching(
+                                Regex(
+                                    ".*Baseline file '$baselinePath' contains 4 reference\\(s\\) to rule ids without a rule set id. For " +
                                         "those references the rule set id 'standard' is assumed. It is advised to regenerate this " +
                                         "baseline file.*",
                                 ),