+++ ktlint-rule-engine-core/src/main/kotlin/com/pinterest/ktlint/rule/engine/core/api/SinceKtlint.kt
@@ -1,0 +1,24 @@
+package com.pinterest.ktlint.rule.engine.core.api
+
+/**
+ * The version in which the rule was introduced.
+ */
+@Repeatable
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.CONSTRUCTOR,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.TYPEALIAS,
+)
+@Retention(AnnotationRetention.BINARY)
+@MustBeDocumented
+public annotation class SinceKtlint(
+    val version: String,
+    val status: Status,
+) {
+    public enum class Status { STABLE, EXPERIMENTAL }
+}