+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/rules/FunctionNamingRule.kt
@@ -3,8 +3,10 @@ package com.pinterest.ktlint.ruleset.standard.rules
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FUN
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.IDENTIFIER
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.IMPORT_DIRECTIVE
-import com.pinterest.ktlint.rule.engine.core.api.Rule
 import com.pinterest.ktlint.rule.engine.core.api.RuleId
+import com.pinterest.ktlint.rule.engine.core.api.SinceKtlint
+import com.pinterest.ktlint.rule.engine.core.api.SinceKtlint.Status.EXPERIMENTAL
+import com.pinterest.ktlint.rule.engine.core.api.SinceKtlint.Status.STABLE
 import com.pinterest.ktlint.ruleset.standard.StandardRule
 import com.pinterest.ktlint.ruleset.standard.rules.internal.regExIgnoringDiacriticsAndStrokesOnLetters
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
@@ -14,9 +16,9 @@ import org.jetbrains.kotlin.psi.KtImportDirective
 /**
  * https://kotlinlang.org/docs/coding-conventions.html#function-names
  */
-public class FunctionNamingRule :
-    StandardRule("function-naming"),
-    Rule.Experimental {
+@SinceKtlint("0.48", EXPERIMENTAL)
+@SinceKtlint("1.0", STABLE)
+public class FunctionNamingRule : StandardRule("function-naming") {
     private var isTestClass = false
 
     override fun beforeVisitChildNodes(